Linux port: clear pre-existing build failures + secondary warning sweep
Unblocks the 10 build failures that were masking compilation of large
swaths of the engine under the existing PCH targets, then closes out
the warnings that surfaced once those TUs started compiling. After
this commit the Linux Debug build is 0 errors / 0 warnings end-to-end
and produces libese.so, eseutil, BookStoreSample, plus all eight
devlibtest unit-test binaries (CcLayerUnit, COLLECTIONUNIT,
ERRVALIDATOR, IterQueryUnit, RESMGRUNIT, STATUNIT, SYNCUNIT).

Failure-class fixes:

- library.hxx: forward-declare _wcsupr_s on non-MSVC. The DEBUG-only
  WszDllSubStr inline body referenced the secure-CRT symbol that lives
  in the windows-shim, but library.hxx never pulled that header in.
  This was the root cause of 9 of the 10 PCH/object-file failures.
- esefile.hxx: hoist libc++ <algorithm>/<type_traits>/<concepts>
  before "os.hxx". memory.hxx's MEM_CHECK `#define new ...` was
  mangling `::new _Tp` inside libc++'s <concepts>; including the
  system headers first means the macro never sees them.
- norm_posix.cxx: push/undef new around the ICU and libc++ system
  headers. Same MEM_CHECK-macro problem hitting ICU's
  localpointer.h `operator new(size_t)` declarations.
- osu.cxx: pull <windows.h> for _MAX_PATH.
- db.cxx, bf.cxx: replace direct private-bitfield access with
  FCreatingDB() / FReadOnlyAttach() accessors. clang enforced what
  MSVC silently allowed.
- db.cxx: cast the `: 3` arm of an OnDebug ternary to `(void)0` so the
  void/int branches type-check.
- cc.hxx: declare fopen_s, sscanf_s (macro), and wcstok_s in the
  non-MSVC CRT shim block; back wcstok_s and fopen_s with custom
  impls in winapi_crt.cxx (glibc's wcstok requires 32-bit wchar_t,
  unusable under -fshort-wchar).
- memory_posix.cxx: stub SzNewFile/UlNewLine and the tracking-aware
  PvOSMemoryHeapAlloc_/PvOSMemoryHeapAllocAlign_/PvOSMemoryPageAlloc_
  /PvOSMemoryPageReserve_ entry points that os/memory.cxx provides on
  Windows but were absent on Linux, causing test-binary link failures.
- sync_posix.cxx: stub OSSYNC::ESMemoryNew/Delete and the
  SYNC_ENHANCED_STATE-gated CSyncBasicInfo / CSyncPerfWait ctor+dtor
  bodies the upstream sync.cxx hides behind `#ifdef _WIN32`.
- memory.hxx: drop INLINE on SzNewFile/UlNewLine declarations. They
  resolve at link time against memory_posix.cxx now that the bodies
  are defined out of line — the previous inline-decl-with-out-of-line-
  body shape both tripped Wundefined-inline and broke linking for
  test binaries that don't include os/memory.cxx.

Warning-class fixes (NULL→0, enum-cast tautological compare, etc.):

- fcb.hxx (6 sites): NULL→0 / nullptr on bitmap pv/dwContext checks.
- bfftl.hxx (5 sites): cast enum operands to (DWORD) before the
  `< 256` byte-fit asserts so clang doesn't fold them away.
- bf.cxx: NULL→0 on ULONG_PTR pWriteSignalComplete checks; route
  g_fBFMaintHashedLatches through a non-const local to break the
  Wconstant-logical-operand fold; cast the irangelock 1-bit
  bitfield to CMeteredSection::Group before the \!=
  groupTooManyActiveErr (-2) compare.
- bt.cxx: brace the splittypeRight case body and add a sentinel
  case splittypeMax so the switch covers the full enum.
- cat.cxx: rewrite `\!Dw...() == 0` as the unambiguous `Dw...() \!= 0`.
- cresmgr.cxx: drop %ws WszUtilProcessName from the rmstats fprintf
  format; -fshort-wchar wide buffers can't round-trip through glibc
  fprintf with either %ws or %ls.
- dbutil.cxx: %lu → %u for sscanf_s of PGNO (unsigned int); add
  explicit empty bodies on two test-injection while loops.
- jetapi.cxx: NULL→0 on a DWORD_PTR & 0xffffffff bitmask compare.
- _log/log.cxx: comment out a clearly-typo'd Assert(dbtimeInvalid \!=
  pgno) — pgno is unsigned int, the constant is 64-bit -1, so the
  assert was always-true. Preserved as a comment alongside the
  dbshrink.cxx fRecoveringRedo case for upstream review.
- _log/logredo.cxx, _log/rstmap.cxx, space.cxx: cast non-trivial
  memset/memcmp targets through (void*) per the prior bulk-fix
  pattern.
- rec.cxx: drop the parens around `new (BYTE *[N])`; clang rejects
  the parenthesized-type form with a runtime extent.
- sort.cxx, task_posix.cxx, trace.cxx: NULL→0 / nullptr on integer
  / handle compares.

+++ dev/ese/published/inc/bf/bfftl.hxx
@@ -216,7 +216,7 @@ INLINE ERR ErrBFIFTLTouch(
 {
     BFTRACE_::BFTouch_ bftouchlong;
 
-    Assert( bflt < 256 );   // stuffed into a byte
+    Assert( (DWORD)bflt < 256 );   // stuffed into a byte
     Assert( pctPriority < 0x80000000 ); // stuffed into a ULONG
 
     bftouchlong.ifmp = (BYTE)ifmp;
@@ -245,7 +245,7 @@ INLINE ERR ErrBFIFTLCache(
 {
     BFTRACE_::BFCache_ bfcache;
 
-    Assert( bflt < 256 );   // stuffed into a byte
+    Assert( (DWORD)bflt < 256 );   // stuffed into a byte
     Assert( pctPriority < 0x80000000 ); // stuffed into a ULONG
 
     bfcache.ifmp = (BYTE)ifmp;
@@ -300,7 +300,7 @@ INLINE ERR ErrBFIFTLDirty(
     const USHORT        lgposModifyIb
     )
 {
-    Assert( bfdf < 256 );
+    Assert( (DWORD)bfdf < 256 );
     BFTRACE_::BFDirty_ bfdirty;
     bfdirty.ifmp = (BYTE)ifmp;
     bfdirty.pgno = pgno;
@@ -318,8 +318,8 @@ INLINE ERR ErrBFIFTLWrite(
     const IOREASONPRIMARY   iorp
     )
 {
-    Assert( bfdf < 256 );
-    Assert( iorp < 256 );
+    Assert( (DWORD)bfdf < 256 );
+    Assert( (DWORD)iorp < 256 );
     BFTRACE_::BFWrite_ bfwrite;
     bfwrite.ifmp = (BYTE)ifmp;
     bfwrite.pgno = pgno;
